Adding user-changeable entry time

+++ src/frontend/common/components/shadcn/calendar.tsx
@@ -36,7 +36,7 @@ function Calendar({
       captionLayout={captionLayout}
       formatters={{
         formatMonthDropdown: (date) =>
-          date.toLocaleString('default', {month: 'short'}),
+          date.toLocaleString('default', {month: 'long'}),
         ...formatters,
       }}
       classNames={{
